Adds encoding to the url before sharing
This will encode the url that is passed to the social platforms.

As the url is used within the sharing url it is important to encode otherwise
if the url contains items such as a hash it will terminate the parsing
of the share url at that stage and miss anything after the hash.

eg.

`https://twitter.com/intent/tweet?url=https://www.ft.com/content/1581cbdb-c47e-3073-b629-017397a981e5#post-44945&text=Example%20Title&related=&via=FT`

When this url is parsed it will think all of this is part of the hash

`#post-44945&text=Example%20Title&related=&via=FT`