Remove the new "scale" variant.
The new scale variant was going to be the same size as the small
variant and scale up on mobile. The reason was for a larger touch
target, but the touch target is already decent because the touch
target is increased beyond the size of the social icon. We might
add it back if we think the touch target could be bigger still,
but for now, no.