Adds descriptive text to links for accessibility
Adds descriptive text to the links when the markup is not provided for the social links and therefore the component is only being used in an enhanced experience manor.

This will improve the accessibility of this component as users of screen readers will hear what the intent of using these links will know that they open in a new window.

Its worth highlighting here that this component has a "link" social option in this
component is an edge case that already behaves  in a possibly unexpected way. Clicking this option will just open the link in a new window, it doesn't do anything like copy to clipboard. For this reason this option needs its own and different descriptive text which is still not overly helpful.

`target="_blank"` and `rel="noopener"` attributes were also added to the anchors as part of this work.